+++ src/vpn/mod.rs
@@ -1,3 +1,4 @@
+pub mod amnezia_share;
 pub mod driver;
 pub mod model;
 pub mod node_api;